API-Fixed code to remove non-active mode from glog
Recent changes in the structure of the data mapping, especially with
`qlabel` being systematically stored, broke routines which expected to
have only quantities related to the data stored in the returned data
object like `__del_nonactive_modes` of `parser.gaussian.glog`.
As a workaround for now, when needed, the iteration simply ignores the
key "qlabel".