Keys used by atomic_data can be int or string
The original implementation returned atomic data as a dictionary with
the atomic symbol as key.  This could create some headache if the
atomic identifiers where originally atomic numbers, since the keys could
not be directly used.  The new code now returns the keys exactly as they
are provided in put to atomic_data.
Since the symbol is not necessarily available anymore, a new subkey has
been created, `symb`.
Routines which accept any kind of atomic labels but need to refer to the
atomic symbol should be updated to use `[key]['symb']`.