Atomic numbers now supported in povray functions
The original versions of the functions relied directly on the
`atomic_data` function to build labels in the PovRay input files as
well, so atomic labels had to be provided as symbols (strings).
After the update of `atomic_data`, the functions have been rewritten to
be more flexible, removing the need for an explicit conversion to atomic
symbols.