Fixed datablock indexes for quantities in glog
For state-specific quantities in Gaussian log files, the system added
a field to read excited-states information (to identify the state of
interest).
This datablock was added before the quantities of interest.
If it was included with the quantity to parse, the system to identify
the indexes containing the actual values could be misled to select
the excited-state information instead of the quantity, resulting in
wrong data read or even false positive (the parser incorrectly expects
the data to be present).