Fixed data type query in fchk.get_hess_dat()
`parser.fchk.get_hess_dat` was not fully updated to the new `QData`
structure and assumed the data type to be a dictionary element of the
data attribute instead of a separate attribute of `QData`.  This caused
some errors when trying to read the Hessian eigenvectors matrix from
a Gaussian formatted checkpoint file.