Fixed parsing of RS/ROA setup in glog for harmonic
The RS/ROA setup in the anharmonic block of the Gaussian output file was
not read for the harmonic approximation while the data were, which
caused a mismatch in the block composition and ESTAMPES getting
confused when parsing Raman/ROA data.
The parser for the setup is now properly added for the harmonic level.