Improvement of the POV-Ray builder library
The code of `visual.povrender` has been refactored, making it more
modular and easy to extend.
A new class, `POVBuilder` has been added to make easier the construction
of description scenes for POV-Ray.  The class should be preferred to
manage the construction of one or more input files.
The construction of input files can be alternatively managed also
through functions.