Ballast: fixed parsing of layout specifications
The previous patch had fixed in part the parsing of the layout block in
INI files but introduced other problems due to a misconstructed exit
condition.  This would cause BALLAST to stop the parsing of layout
parameters after the first one, and in some cases, fail to retrieve the
information to display the curves, resulting in empty blocks.