Added method is_label_set() to SpecLayout
It is now possible to query if the labels for X and Y has been set
through the method is_label_set().
Internally, the instance uses two private attributes to keep track of
the setup.

Note that to make the logic works, if the constructor receives `None`
for xlabel and/or ylabel (the default), the latter are assumed not set.
To deactivate these labels and have them correctly understood as set,
they must be set through their respective setters.