Fixed test to check if atom clicked on molecule
The test in `visual.molview` to check if an atom was clicked did not
check the right event, resulting in the possibility of the click not
being registered on some machines.  It seemed to work on Linux with an
older version of Qt 6 but failed on Mac OS X with a newer version.