Fixed parsing of anh. static Raman from glog file
ESTAMPES did not differentiate static and dynamic Raman when parsing
anharmonic data.  However, there is no setup or frequency information
typically provided for the former, resulting in undefined quantities.
This led ESTAMPES to attempt some division of blocks based on the number
of setups, which was, however, not set.
The parser now differentiates the static and dynamic case.