Fixed support of qty=3xx from fchk files
There were several issues with the previous version of ESTAMPES to
support frequency-dependent properties.
The parsing was too simple and:
- did not make any post-processing to reconstruct lower-triangular
  quantities
- did not try to put into shape the quantity.
- could not process some keywords like "alpha(-w,w)" because of an
  incomplete regular expression.

The new code is more versatile and provide data with a more manageable
form (still pure Python).

As a side effect of the fixes done for property of type 3xx, the parsing
of the dipole and rotatory strengths have been fixed.