Fixed Spectrum.get_[X|Y]unit with no unit
Spectrum.get_xunit and Spectrum.get_yunit never checked if no unit was
set (in which case the internal value is None) and simply proceeded.
This could give an error if a conversion to LaTeX format was requested
since some string operations were carried out.
A similar problem could raise in BALLAST, which did not check if no unit
was available for a given curve.  Now, absent units are simply ignored
and BALLAST assumes that they are consistent with the others.