Ballast now support filetype specifications.
One limit of Ballast is that it was only using the basic features of
DataFile, which relied on the file extension to guess the file type.
This could be misleading in some cases or the file may not be
recognizable due to the use of "exotic" extensions, for instance in
debugging.
The filetype can now be provided to override the automatic setting.
It is possible to provide filetypes for each source file, by separating
them with "&".