convert_y now provides information to correct Y0
The dependency of the final intensity on the X axis is a bit subtler
than originally done.  Some conversions can/must be applied to each
value of the energy (x), while some are intended to correct the original
definition of Y.  The two are not anymore mixed and provided as separate
optional functions by `convert_y`.

`broaden` now uses this information to correctly compute the intensity.