Fixed regression from previous changes to BALLAST
The previous changes introduced the possibility to set up the Raman
incident frequencies and setups in BALLAST.  This led to extra
parameters for the Spectrum class to be systematically set.  However,
the constructor of the class did not consider the case of parameters
created but unset (None).  The constructor has been made more robust.