MolUI can export screenshot (experimental)
Added shortcut to save a screenshot: CTRL+P
As usual, the documentation of Qt is crap and the method to do it was
based on some forums, especially:
https://stackoverflow.com/questions/79174016/render-the-view-of-a-qt3dwindow-to-an-image

The system seems to work overall well but gives completely wrong colors
on Mac OS X Retina.  Other systems have not been tested at the time
of commit.