get_hess_data failed to return evec/eval from GLog
`DataFile.get_hess_data` failed to return the eigenvalues or
eigenvectors from a Gaussian log file if it did not contain the
Cartesian force constants matrix.  This was due to a flawed logic that
made the function bypass such possibility.  The logic has been fixed by
conditioning the forced calculation to the presence of the Cartesian
force constants matrix.