Fixed null division warning in vibrational motions
`visual.povrender` or `visual.VibView` could display a warning message
if the vibrational motion to display had a null component on one atom.
The object creation on this atom is now bypassed.

Note that this could create problems in the future if we want to
operate on individual objects since the list of objects and
transformations is not consistent with the list of atoms.
An alternative could be to switch to dictionaries, not contemplated for
now.