Translation operation in `MolUI` limited to zoom
The original translation vector used for the transformed molecular
coordinates proposed when exporting the geometry in `MolUI` used the raw
translation coordinates from the orbital camera, which rotate around the
center of the view.  This led to huge translational vectors, with little
sense mathematically.  The new model only consider the zoom operation,
as a displacement along Z.