Support of 3rd and 4th energy deriv. from GLog
Added support of 3rd and 4th derivatives of the energy from Gaussian log
files.
However, the data can only be extracted from the VPT2 block for now.
Since this block is partial and may follow a different numbering, the
data are extracted as a dictionary and not a matrix.
Because of these limitations, users should avoid using it as its form
may change in the future, but it may be convenient in some cases for
now, until an improved parser is built for these quantities.