Added temperature weigh model for vib. transitions
ESTAMPES can now compute a weight for each vibrational transition to
simulate the temperature effects assuming that all possible transitions
are included.
The model is valid for the pure harmonic-oscillator approximation and in
part within the simple perturbation theory-like model for VPT2.

The implementation introduces 2 new functions in `tools.vib`:
- `get_vib_trans` to construct a simple dictionary summarizing the
  changes associated to a given transition.
- `weigh_trans_progress` to construct the weigh-related scaling factor
  to the harmonic intensities.